requirement.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
crypto==1.4.1
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Naked==0.1.32
pycryptodome==3.18.0
pypng==0.20220715.0
PyQRCode==1.2.1
requests==2.31.0
shellescape==3.8.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.1
RPi.GPIO==0.7.1
mysql-connector-python==8.1.0